--- a/tests/test-output-pgsql-validgeom.cpp
+++ b/tests/test-output-pgsql-validgeom.cpp
@@ -83,7 +83,7 @@ void test_z_order() {
     db->assert_has_table("osm2pgsql_test_polygon");
     db->assert_has_table("osm2pgsql_test_roads");
 
-    db->check_count(10, "SELECT COUNT(*) FROM osm2pgsql_test_polygon");
+    db->check_count(12, "SELECT COUNT(*) FROM osm2pgsql_test_polygon");
     db->check_count(0, "SELECT COUNT(*) FROM osm2pgsql_test_polygon WHERE NOT ST_IsValid(way)");
     db->check_count(0, "SELECT COUNT(*) FROM osm2pgsql_test_polygon WHERE ST_IsEmpty(way)");
 }