--- a/tests/test-output-pgsql.cpp
+++ b/tests/test-output-pgsql.cpp
@@ -90,7 +90,7 @@ void test_regression_simple() {
     db->check_count(1342, "SELECT count(*) FROM osm2pgsql_test_point");
     db->check_count(3231, "SELECT count(*) FROM osm2pgsql_test_line");
     db->check_count( 375, "SELECT count(*) FROM osm2pgsql_test_roads");
-    db->check_count(4127, "SELECT count(*) FROM osm2pgsql_test_polygon");
+    db->check_count(4136, "SELECT count(*) FROM osm2pgsql_test_polygon");
 
     // Check size of lines
     db->check_number(1696.04, "SELECT ST_Length(way) FROM osm2pgsql_test_line WHERE osm_id = 1101");
@@ -142,7 +142,7 @@ void test_latlong() {
     db->check_count(1342, "SELECT count(*) FROM osm2pgsql_test_point");
     db->check_count(3229, "SELECT count(*) FROM osm2pgsql_test_line");
     db->check_count(374, "SELECT count(*) FROM osm2pgsql_test_roads");
-    db->check_count(4127, "SELECT count(*) FROM osm2pgsql_test_polygon");
+    db->check_count(4136, "SELECT count(*) FROM osm2pgsql_test_polygon");
 
     // Check size of lines
     db->check_number(0.0105343, "SELECT ST_Length(way) FROM osm2pgsql_test_line WHERE osm_id = 1101");
@@ -279,7 +279,7 @@ void test_clone() {
     db->check_count(1342, "SELECT count(*) FROM osm2pgsql_test_point");
     db->check_count(3231, "SELECT count(*) FROM osm2pgsql_test_line");
     db->check_count( 375, "SELECT count(*) FROM osm2pgsql_test_roads");
-    db->check_count(4127, "SELECT count(*) FROM osm2pgsql_test_polygon");
+    db->check_count(4136, "SELECT count(*) FROM osm2pgsql_test_polygon");
 }
 
 } // anonymous namespace