@mmd-osm 2018-08-11
If way processing isn't totally IO bound, there's chances that the single threaded initial processing phase would benefit even more from luajit than it did in the past.