@lonvia 2018-08-18
I had forgotten the part where tags are copied from ways to relations. That is gone now, too. I think that covers all old-style MP code.

@pnorman, it would help if you could look over the tagtransform code, if the handling is still right