@lonvia 2018-09-01
There is no more pending step during import but it is still needed for updates, where changes on nodes/way need to be propagated upwards to ways/relations.