@bagage 2018-09-24
@pnorman Did you have the time to look at this? You seem *really* busy but it'd be awesome to have this I think (when I see how much ways need too be processed for France only... it's killing my server!).