@lonvia 2018-12-12
This has been lying around for a while now. I have some more refactoring changes pending and like to get this in to avoid merge conflicts. If there are no further objections by the end of the week, I'll just get it merged.